requirements_api.txt
# Django and core
Django==5.2.6
asgiref==3.9.2
sqlparse==0.5.3
tzdata==2025.2

# REST API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
drf-spectacular==0.28.0

# Google AI
google-generativeai

# Additional utilities
PyJWT==2.10.1
python-dotenv==1.0.1
